Changed validation rule when sending crawl resp.
We assumed that we are always sending blocks in our own chain when responding to a crawl request. However, the new IPv8 crawler can request blocks that are not in your chain. This check was not consistent with that notion, which results in many missing blocks for the crawler (since they were simply not being sent).