Made should_sign method asynchronous
The should_sign currently returns a boolean value, indicating whether we should sign an incoming block or not. However, it is possible that some asynchronous operation needs to happen before we can determine whether an incoming block should be signed (e.g., fetching some network data). To allow for this use-case, I slightly refactored the should_sign method in the TrustChainCommunity, such that it can deal with listeners that are returning a Deferred in the should_sign method.