Crawling now works without knowing chain length
Our current TrustChain crawling depends on introduction requests that contain an integer with the chain length. However, we cannot always assume that the length of the chain is known when invoking crawl_chain. Therefore, if the latest_block_num argument is None and there is no information about the crawled peer in the database so far, we should first send out a request to fetch the latest block of that peer. By doing so, we find out the chain length and proceed with the crawl.