Binding on_should_sign_outcome vars at runtime
When working on the Noodle protocol, we noticed that if the duration between invocation of the addCallback(..., on_should_sign_outcome) method and calling the on_should_sign_outcome method is long, and process_half_block is invoked again in the meantime, the variables used by the on_should_sign_outcome method are incorrect. Instead, we should immediately bind the blk, peer and validation variables, which is what this PR proposes. Furthermore, I extracted the on_should_sign method from the process_half_block method, which makes the process_half_block method more readable.