Change lambda's parameter name
Changed lambda's parameter name from "sign" to "sig". The lambda uses "sig" while "sign" is how the parameter is named. This creates an undefined reference to "sig", a reference which is fixed with the name change. Since "sig" is used elsewhere in the function the lambda's in, "sign" was changed into "sig" to fix the bug and make it consistent.