Unite sys-path modifications as __scriptpath__.py
Instead of modifying system path on a per-module basis it is instead modified in a __scriptpath__.py file that is imported into the desired module.

In addition instead of appending into the system path the modification inserts into it to prevent path overrides.

Instruct flake8 to ignore import violations

Since the side effects of importing __scriptpath__.py are desired, flake8 is instructed to ignored the violations.