Change bare exceptions. (#655)
* Change bare exceptions

Change bare exceptions to exceptions specified in error report [#650](https://github.com/Tribler/py-ipv8/issues/650).

Added RuntimeError to try-except

Added RuntimeError to try-except since a unit-test detected such an error. Not adding it is a regression from pre-PR because pre-PR did not crash on it.

* Change logger to instance logger

Changed logger to instance logger. Prefixed it with a single underscore.